requirements-full.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
edge-tts==6.1.9
moviepy==1.0.3
pillow==10.1.0
numpy==1.26.2
assemblyai==0.17.0
python-dotenv==1.0.0
gunicorn==21.2.0 